requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.0
tensorflow>=2.8.0
numpy>=1.19.0
psycopg2-binary>=2.9.0
pika>=1.2.0
cryptography>=3.4.0
face-recognition>=1.3.0
Flask>=2.0.0
Flask-CORS>=3.0.0
pyyaml>=5.4.0
python-dotenv>=0.19.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
prometheus-client>=0.14.0
psutil>=5.8.0
gputil>=1.4.0
hvac>=0.11.0
boto3>=1.20.0
httpx>=0.23.0
